% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_binom_gest.R
\name{add_binom_gest}
\alias{add_binom_gest}
\title{Add binomial G-Estimate}
\usage{
add_binom_gest(
  tbl,
  x,
  n,
  cred_level = 0.8,
  theta_from = 0.01,
  theta_to = 0.99,
  theta_by = 0.01
)
}
\arguments{
\item{tbl}{a table which *requires* two columns. A column with the number of
successes (x) and a column with totals (n).}

\item{x}{column with number of successes}

\item{n}{column with totals}

\item{cred_level}{level of credible interval to compute. For example
and interval of .80 will return a 20th percentile lower interval and a
80th percentile upper interval}

\item{theta_from}{start of parameter space}

\item{theta_to}{end of parameter space}

\item{theta_by}{increment of parameter space}
}
\value{
The original table with the following columns
  \item{.gest_dist}{Full posterior distribution. Grain may be increased with the theta_* parameters}
  \item{.raw}{Estimate (success / total)}
  \item{.gest}{Posterior EB g-estimate}
  \item{.lo}{Lower bound of credible interval for EB estimate}
  \item{.hi}{Upper bound of credible interval for EB estimate}
}
\description{
adds an estimated shinkage value obtained by G-Modeling
}
\examples{

library(tibble)
set.seed(2017)

# simulate 200 random examples from a beta-binomial
obs <- 200
dat <- tibble(prob = rbeta(obs, 10, 50),
                  n = round(rlnorm(obs, 4, 2)) + 1,
                  x = rbinom(obs, n, prob))

result <- add_binom_gest(dat, x, n)
result
}
