% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_poisson_gest.R
\name{add_poisson_gest}
\alias{add_poisson_gest}
\title{Add binomial G-Estimate}
\usage{
add_poisson_gest(
  tbl,
  value,
  cred_level = 0.8,
  theta_from = -4,
  theta_to = 5,
  theta_by = 0.1,
  probs = c(0.005, 0.025, 0.165, 0.25, 0.5, 0.75, 0.835, 0.975, 0.995)
)
}
\arguments{
\item{tbl}{a table which *requires* two columns. A column with the number of
successes (x) and a column with totals (n).}

\item{cred_level}{level of credible interval to compute. For example
and interval of .80 will return a 20th percentile lower interval and a
80th percentile upper interval}

\item{theta_from}{start of parameter space}

\item{theta_to}{end of parameter space}

\item{theta_by}{increment of parameter space}

\item{x}{column with number of successes}

\item{n}{column with totals}
}
\value{
The original table with the following columns
  \item{.gest_dist}{Full posterior distribution. Grain may be increased with the theta_* parameters}
  \item{.raw}{Estimate (success / total)}
  \item{.gest}{Posterior EB g-estimate}
  \item{.lo}{Lower bound of credible interval for EB estimate}
  \item{.hi}{Upper bound of credible interval for EB estimate}
}
\description{
adds an estimated shinkage value obtained by G-Modeling
}
\examples{

dontrun{
#Example 1 ##################################
library(tidyverse)
library(ZIM)
data(syph)
df <- syph \%>\%
  pivot_longer(cols = starts_with("a"), names_to = "key", values_to = "value")

df \%>\%
  add_poisson_gest(value)
  }
#Example 2 ##################################
data("df")
df \%>\% add_poisson_gest() \%>\% tidyr::unnest(.probs)
}
